move deploy_kn_k8s_svc_in_same_namespace_test.go
