+++ test/servinge2e/deploy_kn_k8s_svc_in_same_namespace_test.go
@@ -1,4 +1,4 @@
-package kourier
+package servinge2e
 
 import (
 	"context"
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/openshift-knative/serverless-operator/test"
-	"github.com/openshift-knative/serverless-operator/test/servinge2e"
 	routev1 "github.com/openshift/api/route/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -39,7 +38,7 @@ func TestKnativeVersusKubeServicesInOneNamespace(t *testing.T) {
 	}
 
 	// Check Kube service responds
-	servinge2e.WaitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
+	WaitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
 
 	// Deploy Knative service in the same namespace
 	ksvc, err := test.WithServiceReady(caCtx, helloworldService, testNamespace2, image)
@@ -48,14 +47,14 @@ func TestKnativeVersusKubeServicesInOneNamespace(t *testing.T) {
 	}
 
 	// Check that both services respond
-	servinge2e.WaitForRouteServingText(t, caCtx, ksvc.Status.URL.URL(), helloworldText)
-	servinge2e.WaitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
+	WaitForRouteServingText(t, caCtx, ksvc.Status.URL.URL(), helloworldText)
+	WaitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
 
 	// Delete Knative service
 	caCtx.Clients.Serving.ServingV1().Services(testNamespace2).Delete(context.Background(), ksvc.Name, metav1.DeleteOptions{})
 
 	// Check that Kube service still responds
-	servinge2e.WaitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
+	WaitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
 
 	// Remove the Kube service
 	caCtx.Clients.Route.Routes(testNamespace2).Delete(context.Background(), svc.Name, metav1.DeleteOptions{})
@@ -69,7 +68,7 @@ func TestKnativeVersusKubeServicesInOneNamespace(t *testing.T) {
 	}
 
 	// Check that Knative service responds
-	servinge2e.WaitForRouteServingText(t, caCtx, ksvc.Status.URL.URL(), helloworldText)
+	WaitForRouteServingText(t, caCtx, ksvc.Status.URL.URL(), helloworldText)
 
 	//Create deployment
 	err = test.CreateDeployment(caCtx, kubeHelloworldService, testNamespace2, image)
@@ -91,8 +90,8 @@ func TestKnativeVersusKubeServicesInOneNamespace(t *testing.T) {
 	}
 
 	// Check that both services respond
-	servinge2e.WaitForRouteServingText(t, caCtx, ksvc.Status.URL.URL(), helloworldText)
-	servinge2e.WaitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
+	WaitForRouteServingText(t, caCtx, ksvc.Status.URL.URL(), helloworldText)
+	WaitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
 
 	// Remove the Kube service
 	caCtx.Clients.Route.Routes(testNamespace2).Delete(context.Background(), svc.Name, metav1.DeleteOptions{})
@@ -100,7 +99,7 @@ func TestKnativeVersusKubeServicesInOneNamespace(t *testing.T) {
 	caCtx.Clients.Kube.AppsV1().Deployments(testNamespace2).Delete(context.Background(), svc.Name, metav1.DeleteOptions{})
 
 	// Check that Knative service still responds
-	servinge2e.WaitForRouteServingText(t, caCtx, ksvc.Status.URL.URL(), helloworldText)
+	WaitForRouteServingText(t, caCtx, ksvc.Status.URL.URL(), helloworldText)
 
 	// Delete the Knative service
 	caCtx.Clients.Serving.ServingV1().Services(testNamespace2).Delete(context.Background(), ksvc.Name, metav1.DeleteOptions{})
