+++ test/servinge2e/helpers.go
@@ -12,6 +12,16 @@ import (
 	"knative.dev/pkg/test/spoof"
 )
 
+const (
+	testNamespace         = "serverless-tests"
+	testNamespace2        = "serverless-tests2"
+	image                 = "gcr.io/knative-samples/helloworld-go"
+	helloworldService     = "helloworld-go"
+	helloworldService2    = "helloworld-go2"
+	kubeHelloworldService = "kube-helloworld-go"
+	helloworldText        = "Hello World!"
+)
+
 func WaitForRouteServingText(t *testing.T, caCtx *test.Context, routeURL *url.URL, expectedText string) {
 	t.Helper()
 	if _, err := pkgTest.WaitForEndpointState(
