+++ test/servinge2e/kourier/const.go
@@ -1,11 +1,6 @@
 package kourier
 
 const (
-	testNamespace         = "serverless-tests"
-	testNamespace2        = "serverless-tests2"
-	image                 = "gcr.io/knative-samples/helloworld-go"
-	helloworldService     = "helloworld-go"
-	helloworldService2    = "helloworld-go2"
-	kubeHelloworldService = "kube-helloworld-go"
-	helloworldText        = "Hello World!"
+	image          = "gcr.io/knative-samples/helloworld-go"
+	helloworldText = "Hello World!"
 )
